third_party/blink/web_tests/platform/generic/http/tests/security/xss-eval-expected.txt

This page verifies that you can't use eval to subvert cross-domain checks.

If the test passes, you'll see a pass message below.

PASS: eval.call(frames[0], 'document').testExpando should be It's me! and is.
PASS: childEval.call(frames[0], 'document').testExpando should be TypeError and is.
PASS: childEvalCaller('document').testExpando should be SecurityError and is.
PASS: childLocalEvalCaller('document').testExpando should be TypeError and is.
